Export CBR, editable fps, cylinder-carousel template, batch transitions, timing + group fixes
Export:
- Force strict CBR (minrate=maxrate=b:v, nal-hrd=cbr) so the H.264 output
  actually reaches the target bitrate. Capped ABR undershot badly on
  compressible content, so a 600 MB estimate came out ~57 MB and the
  bitrate control barely changed the file. Verified with ffmpeg.

Editing frame rate:
- Comp fps is now editable (Composition Settings) with standard rates
  incl. fractional 23.976/29.97/59.94; fps changed to f32 end-to-end.

Template maker:
- New "Cylinder Carousel" guided dialog + create_cylinder_carousel
  command: images wrap evenly around a cylinder, Y-rotation keyframed to
  snap between them (hold + ease), looping seamlessly; optional per-image
  fade. Choose image count/order and hold/snap speeds.

Transitions:
- Apply in/out transitions to a whole multi-selection at once (layer
  context menu), and draw transition wedges on timeline clips so applied
  transitions are visible.
- Right-click no longer collapses a multi-selection (drag handlers ignore
  non-left buttons), so batch actions see all selected layers.

Timing + groups:
- Inspector "Timing" section sets a layer's (or the whole selection's)
  start/duration numerically.
- set_layer_range now carries a group's children (ranges + keyframes)
  when the group is moved, recursing through nested groups.

+++ src-tauri/src/model.rs
@@ -18,7 +18,7 @@ use crate::text::Font;
 pub struct Project {
     pub width: u32,
     pub height: u32,
-    pub fps: u32,
+    pub fps: f32,
     pub duration_ms: u32,
     /// Drawn back-to-front: index 0 is the bottom layer.
     pub layers: Vec<Layer>,
@@ -1317,7 +1317,7 @@ impl Project {
         Project {
             width: 1920,
             height: 1080,
-            fps: 30,
+            fps: 30.0,
             duration_ms: 4000,
             layers: vec![],
             media: vec![],
@@ -1415,7 +1415,7 @@ impl Project {
         Project {
             width: w,
             height: h,
-            fps: 30,
+            fps: 30.0,
             duration_ms: 4000,
             layers: vec![backdrop, accent, title],
             media: vec![],
